Added simple polygon tessalator

+++ GaitSymQt/MeshStore.cpp
@@ -60,11 +60,11 @@ void MeshStore::addMesh(const MeshStoreObject &meshStoreObject)
 }
 
 void MeshStore::addMesh(const std::string &path, const std::vector<double> &vertexList, const std::vector<double> &normalList,
-                        const std::vector<double> &colourList, const std::vector<double> &uvList, const pgd::Vector3 &lowerBound, const pgd::Vector3 &upperBound)
+                        const std::vector<float> &colourList, const std::vector<double> &uvList, const pgd::Vector3 &lowerBound, const pgd::Vector3 &upperBound)
 {
     if ((vertexList.size() * sizeof(double) +
          normalList.size() * sizeof(double) +
-         colourList.size() * sizeof(double) +
+         colourList.size() * sizeof(float) +
          uvList.size() * sizeof(double) +
          sizeof(MeshStoreObject)) > m_targetMemory) return;
